allocate a render semaphore for each swapchain image
A recent update to the validation layer has highlighted what I think may
be a bug with the Vulkan tutorial. The validation error recommends
allocating a semaphore for each swapchain image when signalling when a
frame has been presented, which makes sense instead of allocating and
indexing based on in flight images.